requirements.txt
aiogram==3.15.0
groq>=0.9.0
SpeechRecognition==3.10.4
pydub==0.25.1
requests==2.32.3
aiohttp==3.10.5
python-dotenv
asyncpg==0.29.0
greenlet==3.0.3
replicate>=0.25.0
Pillow>=10.0.0
supabase>=2.0.0